--- ArduinoMacroKeyboard/src/configuration.txt
@@ -1,6 +1,0 @@
-#pragma once
-#ifndef Configuration_h
-#define Configuration_h
-    #include "Arduino.h"
-
-#endif
